Fixes patrikerdes/gradle-use-latest-versions-plugin#43  * Addressing CodeNarc issues
