+++ src/main/groovy/se/patrikerdes/Common.groovy
@@ -156,13 +156,11 @@ class Common {
     static List<String> getGradleConfigFilesOnPath(String absolutePath) {
         List<String> filePaths = new FileNameFinder().getFileNames(absolutePath, '**/*.gradle')
         filePaths += new FileNameFinder().getFileNames(absolutePath, '**/gradle.properties')
-        return filePaths
     }
 
     static List<String> getKotlinConfigFilesOnPath(String absolutePath) {
         List<String> filePaths = new FileNameFinder().getFileNames(absolutePath, '**/*.gradle.kts')
         filePaths += new FileNameFinder().getFileNames(absolutePath, 'buildSrc/**/*.kt')
-        return filePaths
     }
 
 }
