+++ src/test/groovy/se/patrikerdes/variableupdatetypespecific/VariableUpdatesCommonGradleFunctionalTest.groovy
@@ -1,11 +1,11 @@
-package se.patrikerdes.VariableUpdatesTypeSpecificFunctionalTests
+package se.patrikerdes.variableupdatetypespecific
 
 import org.gradle.testkit.runner.BuildResult
 import se.patrikerdes.BaseFunctionalTest
 import se.patrikerdes.CurrentVersions
 
-class VariableUpdatesGradlePropertiesFunctionalTest extends BaseFunctionalTest {
-    void "a variable assigned in gradle.properties will be updated"() {
+class VariableUpdatesCommonGradleFunctionalTest extends BaseFunctionalTest {
+    void "a variable assigned in test.gradle will be updated"() {
         given:
         buildFile << """
             plugins {
@@ -14,6 +14,7 @@ class VariableUpdatesGradlePropertiesFunctionalTest extends BaseFunctionalTest {
             }
 
             apply plugin: 'java'
+            apply from: "test.gradle"
 
             repositories {
                 mavenCentral()
@@ -24,22 +25,24 @@ class VariableUpdatesGradlePropertiesFunctionalTest extends BaseFunctionalTest {
                 compile "log4j:log4j:\$log4j_version"
             }
         """
-        File gradlePropertiesFile = testProjectDir.newFile('gradle.properties')
-        gradlePropertiesFile << '''
-            junit_version = 4.0
-            log4j_version=1.2.16
+        File commonGradleFile = testProjectDir.newFile('test.gradle')
+        commonGradleFile << '''
+        ext {
+            junit_version = '4.0'
+            log4j_version='1.2.16'
+        }
         '''
 
         when:
         useLatestVersions()
-        String updatedGradlePropertiesFile = gradlePropertiesFile.getText('UTF-8')
+        String updatedCommonGradleFile = commonGradleFile.getText('UTF-8')
 
         then:
-        updatedGradlePropertiesFile.contains("junit_version = $CurrentVersions.JUNIT")
-        updatedGradlePropertiesFile.contains("log4j_version=$CurrentVersions.LOG4J")
+        updatedCommonGradleFile.contains("junit_version = '$CurrentVersions.JUNIT'")
+        updatedCommonGradleFile.contains("log4j_version='$CurrentVersions.LOG4J'")
     }
 
-    void "will update variables in gradle.properties when specified twice in build gradle"() {
+    void "will update variables in test.gradle when specified twice in build gradle"() {
         given:
         buildFile << """
             plugins {
@@ -48,6 +51,7 @@ class VariableUpdatesGradlePropertiesFunctionalTest extends BaseFunctionalTest {
             }
 
             apply plugin: 'java'
+            apply from: "test.gradle"
 
             repositories {
                 mavenCentral()
@@ -59,22 +63,24 @@ class VariableUpdatesGradlePropertiesFunctionalTest extends BaseFunctionalTest {
                 runtimeOnly "log4j:log4j:\$log4j_version"
             }
         """
-        File gradlePropertiesFile = testProjectDir.newFile('gradle.properties')
-        gradlePropertiesFile << '''
-            junit_version = 4.0
-            log4j_version=1.2.16
+        File commonGradleFile = testProjectDir.newFile('test.gradle')
+        commonGradleFile << '''
+        ext {
+            junit_version = '4.0'
+            log4j_version='1.2.16'
+        }
         '''
 
         when:
         useLatestVersions()
-        String updatedGradlePropertiesFile = gradlePropertiesFile.getText('UTF-8')
+        String updatedCommonGradleFile = commonGradleFile.getText('UTF-8')
 
         then:
-        updatedGradlePropertiesFile.contains("junit_version = $CurrentVersions.JUNIT")
-        updatedGradlePropertiesFile.contains("log4j_version=$CurrentVersions.LOG4J")
+        updatedCommonGradleFile.contains("junit_version = '$CurrentVersions.JUNIT'")
+        updatedCommonGradleFile.contains("log4j_version='$CurrentVersions.LOG4J'")
     }
 
-    void "will not update variables in root gradle.properties in Multi-project without --update-root-properties"() {
+    void "will not update variables in root test.gradle in Multi-project without --update-root-properties"() {
         given:
         buildFile << """
             plugins {
@@ -86,15 +92,18 @@ class VariableUpdatesGradlePropertiesFunctionalTest extends BaseFunctionalTest {
                 apply plugin: 'se.patrikerdes.use-latest-versions'
                 apply plugin: 'com.github.ben-manes.versions'
                 apply plugin: 'java'
+                apply from: "\$rootDir/test.gradle"
                 repositories {
                     mavenCentral()
                 }
             }
         """
-        File rootGradlePropertiesFile = testProjectDir.newFile('gradle.properties')
-        rootGradlePropertiesFile << '''
-            junit_version = 4.0
-            log4j_version=1.2.16
+        File rootCommonGradleFile = testProjectDir.newFile('test.gradle')
+        rootCommonGradleFile << '''
+        ext {
+            junit_version = '4.0'
+            log4j_version='1.2.16'
+        }
         '''
         File rootGradleSettingsFile = testProjectDir.newFile('settings.gradle')
         rootGradleSettingsFile << '''
@@ -113,12 +122,12 @@ class VariableUpdatesGradlePropertiesFunctionalTest extends BaseFunctionalTest {
         useLatestVersions()
 
         then:
-        String updatedGradlePropertiesFile = rootGradlePropertiesFile.getText('UTF-8')
-        updatedGradlePropertiesFile.contains('junit_version = 4.0')
-        updatedGradlePropertiesFile.contains('log4j_version=1.2.16')
+        String updatedCommonGradleFile = rootCommonGradleFile.getText('UTF-8')
+        updatedCommonGradleFile.contains('junit_version = \'4.0\'')
+        updatedCommonGradleFile.contains('log4j_version=\'1.2.16\'')
     }
 
-    void "will update variables in root gradle.properties for Multi-project build"() {
+    void "will update variables in root test.gradle for Multi-project build"() {
         given:
         buildFile << """
             plugins {
@@ -130,15 +139,18 @@ class VariableUpdatesGradlePropertiesFunctionalTest extends BaseFunctionalTest {
                 apply plugin: 'se.patrikerdes.use-latest-versions'
                 apply plugin: 'com.github.ben-manes.versions'
                 apply plugin: 'java'
+                apply from: "\$rootDir/test.gradle"
                 repositories {
                     mavenCentral()
                 }
             }
         """
-        File rootGradlePropertiesFile = testProjectDir.newFile('gradle.properties')
-        rootGradlePropertiesFile << '''
-            junit_version = 4.0
-            log4j_version=1.2.16
+        File rootCommonGradleFile = testProjectDir.newFile('test.gradle')
+        rootCommonGradleFile << '''
+        ext {
+            junit_version = '4.0'
+            log4j_version='1.2.16'
+        }
         '''
         File rootGradleSettingsFile = testProjectDir.newFile('settings.gradle')
         rootGradleSettingsFile << '''
@@ -157,12 +169,12 @@ class VariableUpdatesGradlePropertiesFunctionalTest extends BaseFunctionalTest {
         useLatestVersionsUpdatingRootProperties()
 
         then:
-        String updatedGradlePropertiesFile = rootGradlePropertiesFile.getText('UTF-8')
-        updatedGradlePropertiesFile.contains("junit_version = $CurrentVersions.JUNIT")
-        updatedGradlePropertiesFile.contains("log4j_version=$CurrentVersions.LOG4J")
+        String updatedCommonGradleFile = rootCommonGradleFile.getText('UTF-8')
+        updatedCommonGradleFile.contains("junit_version = '$CurrentVersions.JUNIT'")
+        updatedCommonGradleFile.contains("log4j_version='$CurrentVersions.LOG4J'")
     }
 
-    void "will update variables in root gradle.properties for Multi-project when plugin applied to subproject only"() {
+    void "will update variables in root test.gradle for Multi-project when plugin applied to subproject only"() {
         given:
         buildFile << """
             plugins {
@@ -174,15 +186,18 @@ class VariableUpdatesGradlePropertiesFunctionalTest extends BaseFunctionalTest {
                 apply plugin: 'se.patrikerdes.use-latest-versions'
                 apply plugin: 'com.github.ben-manes.versions'
                 apply plugin: 'java'
+                apply from: "\$rootDir/test.gradle"
                 repositories {
                     mavenCentral()
                 }
             }
         """
-        File rootGradlePropertiesFile = testProjectDir.newFile('gradle.properties')
-        rootGradlePropertiesFile << '''
-            junit_version = 4.0
-            log4j_version=1.2.16
+        File rootCommonGradleFile = testProjectDir.newFile('test.gradle')
+        rootCommonGradleFile << '''
+        ext {
+            junit_version = '4.0'
+            log4j_version='1.2.16'
+        }
         '''
         File rootGradleSettingsFile = testProjectDir.newFile('settings.gradle')
         rootGradleSettingsFile << '''
@@ -201,12 +216,12 @@ class VariableUpdatesGradlePropertiesFunctionalTest extends BaseFunctionalTest {
         useLatestVersionsUpdatingRootProperties()
 
         then:
-        String updatedGradlePropertiesFile = rootGradlePropertiesFile.getText('UTF-8')
-        updatedGradlePropertiesFile.contains("junit_version = $CurrentVersions.JUNIT")
-        updatedGradlePropertiesFile.contains("log4j_version=$CurrentVersions.LOG4J")
+        String updatedCommonGradleFile = rootCommonGradleFile.getText('UTF-8')
+        updatedCommonGradleFile.contains("junit_version = '$CurrentVersions.JUNIT'")
+        updatedCommonGradleFile.contains("log4j_version='$CurrentVersions.LOG4J'")
     }
 
-    void "will update variables in root gradle.properties for Multi-project when present in multiple projects"() {
+    void "will update variables in root test.gradle for Multi-project when present in multiple projects"() {
         given:
         buildFile << """
             plugins {
@@ -218,15 +233,18 @@ class VariableUpdatesGradlePropertiesFunctionalTest extends BaseFunctionalTest {
                 apply plugin: 'se.patrikerdes.use-latest-versions'
                 apply plugin: 'com.github.ben-manes.versions'
                 apply plugin: 'java'
+                apply from: "\$rootDir/test.gradle"
                 repositories {
                     mavenCentral()
                 }
             }
         """
-        File rootGradlePropertiesFile = testProjectDir.newFile('gradle.properties')
-        rootGradlePropertiesFile << '''
-            junit_version = 4.0
-            log4j_version=1.2.16
+        File rootCommonGradleFile = testProjectDir.newFile('test.gradle')
+        rootCommonGradleFile << '''
+        ext {
+            junit_version = '4.0'
+            log4j_version='1.2.16'
+        }
         '''
         File rootGradleSettingsFile = testProjectDir.newFile('settings.gradle')
         rootGradleSettingsFile << '''
@@ -254,12 +272,12 @@ class VariableUpdatesGradlePropertiesFunctionalTest extends BaseFunctionalTest {
         useLatestVersionsUpdatingRootProperties()
 
         then:
-        String updatedGradlePropertiesFile = rootGradlePropertiesFile.getText('UTF-8')
-        updatedGradlePropertiesFile.contains("junit_version = $CurrentVersions.JUNIT")
-        updatedGradlePropertiesFile.contains("log4j_version=$CurrentVersions.LOG4J")
+        String updatedCommonGradleFile = rootCommonGradleFile.getText('UTF-8')
+        updatedCommonGradleFile.contains("junit_version = '$CurrentVersions.JUNIT'")
+        updatedCommonGradleFile.contains("log4j_version='$CurrentVersions.LOG4J'")
     }
 
-    void "will not update variables in root gradle.properties for Multi-project when not resolved to same version"() {
+    void "will not update variables in root test.gradle for Multi-project when not resolved to same version"() {
         given:
         buildFile << """
             plugins {
@@ -271,15 +289,18 @@ class VariableUpdatesGradlePropertiesFunctionalTest extends BaseFunctionalTest {
                 apply plugin: 'se.patrikerdes.use-latest-versions'
                 apply plugin: 'com.github.ben-manes.versions'
                 apply plugin: 'java'
+                apply from: "\$rootDir/test.gradle"
                 repositories {
                     mavenCentral()
                 }
             }
         """
-        File rootGradlePropertiesFile = testProjectDir.newFile('gradle.properties')
-        rootGradlePropertiesFile << '''
-            junit_version = 4.0
-            log4j_version=1.2.16
+        File rootCommonGradleFile = testProjectDir.newFile('test.gradle')
+        rootCommonGradleFile << '''
+        ext {
+            junit_version = '4.0'
+            log4j_version='1.2.16'
+        }
         '''
         File rootGradleSettingsFile = testProjectDir.newFile('settings.gradle')
         rootGradleSettingsFile << '''
@@ -305,8 +326,8 @@ class VariableUpdatesGradlePropertiesFunctionalTest extends BaseFunctionalTest {
         BuildResult result = useLatestVersionsUpdatingRootProperties()
 
         then:
-        String updatedGradlePropertiesFile = rootGradlePropertiesFile.getText('UTF-8')
-        updatedGradlePropertiesFile.contains('junit_version = 4.0')
+        String updatedCommonGradleFile = rootCommonGradleFile.getText('UTF-8')
+        updatedCommonGradleFile.contains('junit_version = \'4.0\'')
         result.output.contains("A problem was detected: the variable 'junit_version' has different updated versions " +
                 "in different projects.\nNew updated versions are: '$CurrentVersions.JUNIT' and " +
                 "'$CurrentVersions.JUNIT_DEPS', root config file value won't be be changed.")
